Patch global.json before running dotnet format
- Patch `global.json` before running `dotnet format` to use the same .NET SDK to upgrade the project.
- Specify the exact .NET SDK version when patching global.json, rather than whatever happens to be there already.
