+++ src/DotNetBumper.Core/Upgraders/PackageVersionUpgrader.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
 using System.Text.Json;
-using System.Text.Json.Nodes;
 using MartinCostello.DotNetBumper.Logging;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
@@ -43,7 +42,7 @@ internal sealed partial class PackageVersionUpgrader(
 
             context.Status = StatusMessage($"Updating {name}...");
 
-            using (await TryPatchGlobalJsonAsync(project, upgrade.SdkVersion.IsPrerelease, cancellationToken))
+            using (await PatchedGlobalJsonFile.TryPatchAsync(project, upgrade.SdkVersion, cancellationToken))
             {
                 if (HasDotNetToolManifest(project))
                 {
@@ -67,32 +66,6 @@ internal sealed partial class PackageVersionUpgrader(
         }
 
         return result;
-    }
-
-    private static async Task<PatchedGlobalJsonFile?> TryPatchGlobalJsonAsync(
-        string path,
-        bool isPrerelease,
-        CancellationToken cancellationToken)
-    {
-        var globalJson = FileHelpers.FindFileInProject(path, WellKnownFileNames.GlobalJson);
-
-        if (globalJson != null)
-        {
-            var patched = new PatchedGlobalJsonFile(globalJson);
-
-            try
-            {
-                await patched.TryRemoveSdkVersionAsync(isPrerelease, cancellationToken);
-                return patched;
-            }
-            catch (Exception)
-            {
-                patched.Dispose();
-                throw;
-            }
-        }
-
-        return null;
     }
 
     private static bool HasDotNetToolManifest(string path)
@@ -280,63 +253,5 @@ internal sealed partial class PackageVersionUpgrader(
             Level = LogLevel.Warning,
             Message = "Unable to restore .NET tools for {Directory}.")]
         public static partial void UnableToRestoreTools(ILogger logger, string directory);
-    }
-
-    private sealed class PatchedGlobalJsonFile : IDisposable
-    {
-        private readonly string _filePath;
-        private readonly string _backupPath;
-
-        public PatchedGlobalJsonFile(string source)
-        {
-            _filePath = source;
-            _backupPath = $"{source}.{Guid.NewGuid().ToString()[0..8]}.tmp";
-            File.Copy(_filePath, _backupPath, overwrite: true);
-        }
-
-        public void Dispose()
-            => File.Move(_backupPath, _filePath, overwrite: true);
-
-        public async Task TryRemoveSdkVersionAsync(bool isPrerelease, CancellationToken cancellationToken)
-        {
-            if (!JsonHelpers.TryLoadObject(_filePath, out var globalJson))
-            {
-                return;
-            }
-
-            const string AllowPrereleaseProperty = "allowPrerelease";
-            const string SdkProperty = "sdk";
-            const string VersionProperty = "version";
-
-            // Drop the version from the SDK property in the global.json file
-            // but keep any other content, such as versions for MSBuild SDKs.
-            if (globalJson.TryGetPropertyValue(SdkProperty, out var property) &&
-                property?.GetValueKind() is JsonValueKind.Object)
-            {
-                var sdk = property.AsObject();
-
-                var edited = false;
-
-                if (sdk.TryGetPropertyValue(VersionProperty, out var version) &&
-                    version?.GetValueKind() is JsonValueKind.String)
-                {
-                    edited = sdk.Remove(VersionProperty);
-                }
-
-                if (isPrerelease &&
-                    (!sdk.TryGetPropertyValue(AllowPrereleaseProperty, out var allowPrerelease) ||
-                     allowPrerelease?.GetValueKind() is not JsonValueKind.True))
-                {
-                    sdk.Remove(AllowPrereleaseProperty);
-                    sdk.Add(new(AllowPrereleaseProperty, JsonValue.Create(true)));
-                    edited = true;
-                }
-
-                if (edited)
-                {
-                    await globalJson.SaveAsync(_filePath, cancellationToken);
-                }
-            }
-        }
     }
 }
